JMA GCC 3.4 Compatibility
git-svn-id: https://svn.zsnes.com/rep/zsnes/trunk@1852 8cdf4848-9e13-0410-98e6-c9478ff1fb58

+++ src/jma/jma.cpp
@@ -70,7 +70,7 @@ namespace JMA
     }
     
     //Seek to before file block so we can read the file block
-    stream.seekg(-(file_block_size+UINT_SIZE),ios::end);
+    stream.seekg(-((int)file_block_size+UINT_SIZE),ios::end);
     
     jma_file_info file_info;
     char byte;
